refactor: rename permission-level.enum
