+++ src/invitations/dto/create-invitation-req.dto.ts
@@ -1,5 +1,5 @@
 import { NamespaceRole } from 'src/namespaces/entities/namespace-member.entity';
-import { ResourcePermission } from 'src/permissions/permission-level.enum';
+import { ResourcePermission } from 'src/permissions/resource-permission.enum';
 
 export class CreateInvitationReqDto {
   namespaceRole?: NamespaceRole;
