+++ src/invitations/dto/invitation.dto.ts
@@ -1,6 +1,6 @@
 import { GroupDto } from 'src/groups/dto/group.dto';
 import { NamespaceRole } from 'src/namespaces/entities/namespace-member.entity';
-import { ResourcePermission } from 'src/permissions/permission-level.enum';
+import { ResourcePermission } from 'src/permissions/resource-permission.enum';
 
 export class InvitationDto {
   id: string;
