+++ src/invitations/entities/invitation.entity.ts
@@ -1,6 +1,6 @@
 import { Base } from 'src/common/base.entity';
 import generateId from 'src/utils/generate-id';
-import { ResourcePermission } from 'src/permissions/permission-level.enum';
+import { ResourcePermission } from 'src/permissions/resource-permission.enum';
 import { Column, Entity, Index, PrimaryColumn, BeforeInsert } from 'typeorm';
 import { NamespaceRole } from 'src/namespaces/entities/namespace-member.entity';
