+++ src/invitations/invitations.service.ts
@@ -5,7 +5,7 @@ import { FindOptionsWhere, IsNull, Not, Repository } from 'typeorm';
 import { InvitationDto } from './dto/invitation.dto';
 import { CreateInvitationReqDto } from './dto/create-invitation-req.dto';
 import { NamespaceRole } from 'src/namespaces/entities/namespace-member.entity';
-import { ResourcePermission } from 'src/permissions/permission-level.enum';
+import { ResourcePermission } from 'src/permissions/resource-permission.enum';
 import { AuthService } from 'src/auth/auth.service';
 import { UserInvitationDto as AuthInvitationDto } from 'src/auth/dto/invitation.dto';
 import { GroupsService } from '../groups/groups.service';
