+++ src/permissions/dto/permission.dto.ts
@@ -1,6 +1,6 @@
 import { Expose } from 'class-transformer';
 import { IsEnum, IsNotEmpty } from 'class-validator';
-import { ResourcePermission } from '../permission-level.enum';
+import { ResourcePermission } from '../resource-permission.enum';
 
 @Expose()
 export class PermissionDto {
