+++ src/resources/resources.controller.ts
@@ -5,7 +5,7 @@ import { ResourcesService } from 'src/resources/resources.service';
 import { CreateResourceDto } from 'src/resources/dto/create-resource.dto';
 import { UpdateResourceDto } from 'src/resources/dto/update-resource.dto';
 import { PermissionsService } from 'src/permissions/permissions.service';
-import { ResourcePermission } from 'src/permissions/permission-level.enum';
+import { ResourcePermission } from 'src/permissions/resource-permission.enum';
 import {
   Req,
   Res,
